pyproject.toml
[project]
name = "portfolio-calculation"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "playwright>=1.40.0",
    "openpyxl>=3.1.0",
]
